Rename files to use hyphens instead of capital letters
