+++ src/webextension/background/index.js
@@ -3,7 +3,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 import datastore from "./datastore";
-import initializeMessagePorts from "./messagePorts";
+import initializeMessagePorts from "./message-ports";
 
 // XXX: For now, initialize the datastore on startup and then hook up the
 // button. Eventually, we'll have UX to create new datastores (and persist
