+++ src/webextension/manage/components/item-details.css
