+++ src/webextension/manage/components/item-list.css
