+++ src/webextension/manage/containers/add-item.js
