+++ src/webextension/manage/containers/current-item.css
