+++ src/webextension/manage/message-ports.js
