+++ test/background/message-ports-test.js
@@ -14,7 +14,7 @@ chai.use(sinonChai);
 
 import datastore from "../../src/webextension/background/datastore";
 import initializeMessagePorts from
-       "../../src/webextension/background/messagePorts";
+       "../../src/webextension/background/message-ports";
 
 describe("message ports (background side)", () => {
   let itemId, selfMessagePort, otherMessagePort, selfListener, otherListener;
