+++ test/manage/containers/current-item-test.js
@@ -13,9 +13,9 @@ import thunk from "redux-thunk";
 import { initialState, filledState } from "../mock-redux-state";
 import mountWithL10n from "../../mock-l10n";
 import ItemDetails from
-       "../../../src/webextension/manage/components/itemDetails";
+       "../../../src/webextension/manage/components/item-details";
 import CurrentItem from
-       "../../../src/webextension/manage/containers/currentItem";
+       "../../../src/webextension/manage/containers/current-item";
 import * as actions from "../../../src/webextension/manage/actions";
