+++ test/manage/message-ports-test.js
@@ -10,7 +10,8 @@ import thunk from "redux-thunk";
 
 import { initialState } from "./mock-redux-state";
 import * as actions from "../../src/webextension/manage/actions";
-import initializeMessagePorts from "../../src/webextension/manage/messagePorts";
+import initializeMessagePorts from
+       "../../src/webextension/manage/message-ports";
 
 const middlewares = [thunk];
 const mockStore = configureStore(middlewares);
